UNIFIED DUO
Duo's code has been unified. The animation features of the duo are now
built into the script itself.

Duo's code has also been cleaned up and lifecycle is now similar to it's
SOLO counterpart.

Also, indexDUO.html and indexSOLO.html files are clean and closer to be
used as templates to begin working with the library.