Canvas Resolution Independance
Finally, canvas do not need to be a fixed size when in the HTML. It is
asumed that these apps will run in a FULL SCREEN mode, since it is size
by size stimuli. Therefore, the size of the canvas is automatically
calculated, making the development process faster and easier.